امکان ثبت اعداد اعشار و کیبورد شامل ربع و نیم و ضرایب 10 هم باشد

+++ TeleZanbil/Models/ZanbilItem.cs
@@ -6,7 +6,7 @@ namespace ir.EmIT.TeleZanbil.Models
     {
         public int ZanbilItemId { get; set; }
         public string ItemTitle { get; set; }
-        public float ItemAmount { get; set; }
+        public double ItemAmount { get; set; }
         public virtual Unit ItemUnit { get; set; }
         public virtual Zanbil Zanbil { get; set; }
         public bool IsBought { get; set; }
